refactor(tooling): extract shared sample-app vite config and tsconfig base
The vanilla and mobx apps duplicated a 65-line vite.config.ts (identical
except ports) and byte-identical tsconfig.json. Both now consume
sample-app-shared/tooling: a sampleAppViteConfig factory (ports + the
app's import.meta.url, which anchors static-copy's stripBase to that
app's vite root) and a tsconfig.app.json base; per-app files shrink to
what is genuinely per-app.

Refs #288
