+++ apps/sample-app-lit-mobx/package.json
@@ -28,8 +28,6 @@
     "oxfmt": "catalog:",
     "typescript": "catalog:typescript6-compat",
     "vite": "catalog:",
-    "vite-plugin-checker": "catalog:",
-    "vite-plugin-static-copy": "catalog:",
     "wrangler": "catalog:"
   }
 }
