+++ apps/sample-app-lit-mobx/tsconfig.json
@@ -1,10 +1,5 @@
 {
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "allowJs": true,
-    "types": ["vite/client", "dom-navigation"],
-    "noEmit": true
-  },
+  "extends": "../sample-app-shared/tooling/tsconfig.app.json",
   "include": ["src/**/*", "vite.config.ts"],
   "exclude": ["node_modules", "dist"]
 }
