+++ apps/sample-app-lit-mobx/turbo.json
@@ -9,6 +9,15 @@
         "VITE_SAMPLE_APP_PUSH_STATE"
       ],
       "dependsOn": ["^build", "^transit"]
+    },
+    "typecheck": {
+      // tsconfig extends + the vite factory live in sample-app-shared/tooling,
+      // outside this package's default inputs; ^build:types doesn't hash it
+      "inputs": [
+        "$TURBO_DEFAULT$",
+        "$TURBO_ROOT$/tsconfig.base.json",
+        "$TURBO_ROOT$/apps/sample-app-shared/tooling/**"
+      ]
     }
   }
 }
