+++ apps/sample-app-shared/tooling/tsconfig.app.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../../tsconfig.base.json",
+  "compilerOptions": {
+    "allowJs": true,
+    "types": ["vite/client", "dom-navigation"],
+    "noEmit": true
+  }
+}
